Update.
git-svn-id: file:///home/ryo_on/repos/gxemul-repo/gxemul-legacy/trunk@3219 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/cpu_alpha.c
@@ -25,7 +25,7 @@
  *  SUCH DAMAGE.
  *
  *
- *  $Id: cpu_alpha.c,v 1.14 2005-07-18 11:28:32 debug Exp $
+ *  $Id: cpu_alpha.c,v 1.15 2005-07-18 22:00:06 debug Exp $
  *
  *  Alpha CPU emulation.
  *
@@ -721,7 +721,7 @@ void alpha_pc_to_pointers(struct cpu *cpu)
 {
 	uint64_t cached_pc, physaddr;
 	uint32_t physpage_ofs;
-	int pagenr, table_index;
+	uint64_t pagenr, table_index;
 	uint32_t *physpage_entryp;
 	struct alpha_tc_physpage *ppp;
